notifying PopStateListener on pushState/replaceState + refactoring NavManager
